+++ src/beta/client.rs
@@ -6,23 +6,28 @@
 use std::sync::Arc;
 use crate::client::MunaClient;
 use crate::services::{PredictionService, PredictorService};
-use super::remote::BetaPredictionService;
+use super::openai::OpenAIClient;
+use super::remote::{BetaPredictionService, RemotePredictionService};
 
 /// Client for incubating features.
 #[derive(Clone)]
 pub struct BetaClient {
     /// Make remote predictions.
-    pub predictions: BetaPredictionService
+    pub predictions: BetaPredictionService,
+    /// OpenAI-compatible client.
+    pub openai: OpenAIClient,
 }
 
 impl BetaClient {
 
     pub fn new(
         client: Arc<MunaClient>,
-        _predictors: PredictorService,
-        _predictions: PredictionService,
+        predictors: PredictorService,
+        predictions: PredictionService,
     ) -> Self {
-        let beta_predictions = BetaPredictionService::new(client);
-        Self { predictions: beta_predictions }
+        let beta_predictions = BetaPredictionService::new(client.clone());
+        let remote_predictions = RemotePredictionService::new(client);
+        let openai = OpenAIClient::new(predictors, predictions, remote_predictions);
+        Self { predictions: beta_predictions, openai }
     }
 }